Make the Jupyter report dependencies optional
Fixes #426.

Only the HTML report needs Jupyter, but ipython, ipykernel, nbconvert and
lxml_html_clean were unconditional install requirements. They pull in 44 of
the 113 packages a full install produces, so every user paid for them
whether or not they ever generated a report.

They now live in an optional extra:

    pip install cellbender           # count matrices, summary PDF
    pip install cellbender[report]   # adds the HTML report

The reason they could not simply be moved is that run.py imported
report.py at module scope, and report.py imports IPython at module scope
because it is also the module executed inside report.ipynb. So importing
the pipeline at all required IPython.

report.py is therefore split by role rather than by topic:

- summary.py is what the pipeline itself calls: the summary PDF
  (plot_summary, pca_2d) and the notebook driver (run_notebook_make_html).
  It must stay importable with only the base dependencies.
- report.py keeps everything executed inside the notebook, where IPython is
  present by definition, and imports pca_2d from summary.

Attempting to generate a report without the extra now fails with a message
naming the install command, rather than an opaque subprocess error.

The Makefile, CI and Docker image all install the extra, so developer and
container behaviour is unchanged.

tests/test_summary.py asserts in a subprocess that importing the pipeline
does not pull in IPython, nbconvert or ipykernel, so the coupling cannot
quietly come back.
