+++ README.rst
@@ -63,6 +63,17 @@ As of CellBender v0.4.0, CellBender is compatible with python 3.10+, and can be 
 (and we recommend installing in its own ``conda`` environment to prevent
 conflicts with other software).
 
+The HTML report is generated by executing a Jupyter notebook, which pulls in a
+large dependency tree. Those packages are therefore optional. To generate
+reports, install
+
+.. code-block:: console
+
+  $ pip install cellbender[report]
+
+Everything else, including the count matrices and the summary PDF, is produced
+by the base install.
+
 CellBender is run as a command-line tool, as in
 
 .. code-block:: console
